Introduce Rate Limiting on Requests
Considering this is a self-serve API, the idea is to keep it has safe and long-lasting as possible. Here we collect the ipAddress of the API User along with our configured Limit Request Number name and store in a MemCache to rate limit the API. I think 1 second is fine for the singular requests and 2 seconds for the bulk request seems OK considering sqlites capabilities to be queried.