Adding Basic Auth to the API
I was wanting to authenticate some of the privileged requests to the database and thought HMACSHA1 generator would be suitable for salting and hashing the auth details, for now. Usually I use SHA-2 in other programs, but I think this is fine. Some of this code is bad and needs refactoring, but it does the job for now. Details of use are included in the README, but this commit also adds a new controller to interact with the new table to create privledged user details. I've rate limited the request for 60 seconds, which is high but should stop any corruption from happening. With this commit we are only allowing one privileged user at a time, which makes me think - we'll need to disable root interation with the database. Anyone who interacts with the database at root would be privledged, I guess, but I've also always hidden the database away from the main directory, so that is another task.